added tests for several classes and functions in util

+++ code/branches/testing/test/util/MathTest.cc
@@ -323,7 +323,7 @@ namespace orxonox
     ///////////////////////
     TEST(getUniqueNumber, Test)
     {
-        std::set<float> numbers;
+        std::set<unsigned long> numbers;
 
         const size_t iterations = 1000;
